# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BenchmarkResponse": {
        "properties": {
          "completion_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Tokens"
          },
          "config_fingerprint": {
            "title": "Config Fingerprint",
            "type": "string"
          },
          "config_revision": {
            "title": "Config Revision",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "max_tokens": {
            "title": "Max Tokens",
            "type": "integer"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "output_preview": {
            "title": "Output Preview",
            "type": "string"
          },
          "peak_observed_gpu_memory_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Observed Gpu Memory Bytes"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "proxy_name": {
            "title": "Proxy Name",
            "type": "string"
          },
          "time_to_first_token_ms": {
            "title": "Time To First Token Ms",
            "type": "number"
          },
          "tokens_per_second": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens Per Second"
          },
          "total_seconds": {
            "title": "Total Seconds",
            "type": "number"
          }
        },
        "required": [
          "id",
          "created_at",
          "model_id",
          "proxy_name",
          "config_revision",
          "config_fingerprint",
          "prompt",
          "max_tokens",
          "time_to_first_token_ms",
          "total_seconds",
          "completion_tokens",
          "tokens_per_second",
          "peak_observed_gpu_memory_bytes",
          "output_preview"
        ],
        "title": "BenchmarkResponse",
        "type": "object"
      },
      "BenchmarkRunBody": {
        "properties": {
          "max_tokens": {
            "default": 64,
            "maximum": 512.0,
            "minimum": 1.0,
            "title": "Max Tokens",
            "type": "integer"
          },
          "model_id": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Model Id",
            "type": "string"
          },
          "prompt": {
            "default": "Reply with a short description of the sky.",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          }
        },
        "required": [
          "model_id"
        ],
        "title": "BenchmarkRunBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Web UI for managing llama.cpp models and versions",
    "title": "llama.cpp Docker Manager",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/benchmarks/run": {
      "post": {
        "operationId": "run_benchmark_api_benchmarks_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BenchmarkRunBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Benchmark",
        "tags": [
          "benchmarks"
        ]
      }
    },
    "/api/benchmarks/{model_id}": {
      "get": {
        "operationId": "benchmark_history_api_benchmarks__model_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BenchmarkResponse"
                  },
                  "title": "Response Benchmark History Api Benchmarks  Model Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Benchmark History",
        "tags": [
          "benchmarks"
        ]
      }
    },
...
  }
}